backend/requirements.txt
Django==3.2.25
djangorestframework==3.14.0
requests==2.31.0
urllib3==1.26.18
django-filter==23.1
djangorestframework-simplejwt==5.3.0
